+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 ## [Unreleased]
 ### Added
 - IssuingBillingInvoice resource
+- IssuingBillingTransaction resource
 - installmentCount attribute to IssuingPurchase resource
 
 ## [0.21.0] - 2026-05-28